requirements.txt
pika==1.3.1
kombu==5.3.1
python-decouple==3.8